basics_migrated

+++ main/helpers/timer.cpp
@@ -1,7 +1,7 @@
 #include "timer.h"
 #include "freertos/task.h"
 
-void delay_ms(int64_t milliseconds) {
+void delay_ms(uint32_t milliseconds) {
     vTaskDelay(milliseconds / portTICK_PERIOD_MS);
 }
